Restructured and fix bugs in API

+++ src/modules/customers/useCases/DeleteCustomerUseCase.ts
@@ -1,5 +1,7 @@
 import { inject, injectable } from 'tsyringe';
+
 import { AppError } from '../../../errors/AppError';
+import { IOrdersRepository } from '../../orders/repositories/IOrdersRepository';
 import { ICustomersRepository } from '../repositories/ICustomersRepository';
 
 @injectable()
@@ -7,6 +9,8 @@ export class DeleteCustomerUseCase {
   constructor(
     @inject('CustomersRepository')
     private customersRepository: ICustomersRepository,
+    @inject('OrdersRepository')
+    private ordersRepository: IOrdersRepository,
   ) {}
 
   public async execute(id: number): Promise<boolean> {
@@ -14,6 +18,11 @@ export class DeleteCustomerUseCase {
 
     if (!findCustomer) {
       throw new AppError('Customer not found!', 404);
+    }
+
+    const findOrder = await this.ordersRepository.findByCustomer(findCustomer.id);
+    if (findOrder) {
+      throw new AppError('The customer has orders and cannot be deleted');
     }
     return !!(await this.customersRepository.delete(id));
   }
